LPS-154326 Move submission validation to remote service
